+++ packages/icons/src/ChecksIcon.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 
-export interface UploadIconProps extends React.SVGProps<SVGSVGElement> {
+export interface ChecksIconProps extends React.SVGProps<SVGSVGElement> {
   size?: "sm" | "md" | "lg";
   color?: string;
   className?: string;
@@ -12,7 +12,7 @@ const sizeMap = {
   lg: 32,
 };
 
-const UploadIcon: React.FC<UploadIconProps> = ({
+export const ChecksIcon: React.FC<ChecksIconProps> = ({
   size = "md",
   color = "currentColor",
   className = "",
@@ -24,19 +24,20 @@ const UploadIcon: React.FC<UploadIconProps> = ({
       xmlns="http://www.w3.org/2000/svg"
       width={dimension}
       height={dimension}
-      viewBox="0 0 48 48"
+      viewBox="0 0 24 24"
       fill="none"
       stroke={color}
       strokeWidth={2}
       strokeLinecap="round"
       strokeLinejoin="round"
-      className={`icon icon-tabler icon-tabler-upload ${className}`.trim()}
+      className={`icon icon-tabler icons-tabler-outline icon-tabler-checks ${className}`.trim()}
       {...props}
     >
-      <path d="M28 8H12a4 4 0 00-4 4v20m32-12v8m0 0v8a4 4 0 01-4 4H12a4 4 0 01-4-4v-4m32-4l-3.172-3.172a4 4 0 00-5.656 0L28 28M8 32l9.172-9.172a4 4 0 015.656 0L28 28m0 0l4 4m4-24h8m-4-4v8m-12 4h.02" />
+      <path stroke="none" d="M0 0h24v24H0z" fill="none" />
+      <path d="M7 12l5 5l10 -10" />
+      <path d="M2 12l5 5m5 -5l5 -5" />
     </svg>
   );
 };
 
-UploadIcon.displayName = "UploadIcon";
-export default UploadIcon;
+ChecksIcon.displayName = "ChecksIcon";