+++ packages/drawer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @rubenpazch/drawer
 
+## 2.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @rubenpazch/icons@3.0.2
+
 ## 2.0.1
 
 ### Patch Changes