+++ packages/icons/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @rubenpazch/icons
 
+## 3.0.2
+
+### Patch Changes
+
+- updates
+
 ## 3.0.1
 
 ### Patch Changes